+++ CHANGELOG.md
@@ -21,6 +21,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   is applied to a query.
 - Add the `diagnostics` option, which enables checks about how Flop is used. It
   can only be set in the application environment.
+- Allow setting `max_filters` when deriving `Flop.Schema`.
 
 ### Changed
 
@@ -33,9 +34,14 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   disable.
 - Accept cursor pagination parameters without an order field unless the
   tiebreaker is disabled.
-- Add `tiebreaker/1` and `primary_key/1` to the `Flop.Schema` protocol.
+- Add `tiebreaker/1`, `primary_key/1` and `max_filters/1` to the `Flop.Schema`
+  protocol.
 - Only log a warning about a query that already has an `ORDER BY` clause when
   the `diagnostics` option is enabled in the application environment.
+- Allow to pass `filterable` as an option to query functions to narrow the
+  schema's filterable fields.
+- Apply the `sortable` option passed to a query function by narrowing the
+  schema's list instead of replacing it.
 - Raise an `ArgumentError` when a `Flop` struct with a non-matching pagination
   type is passed to `Flop.to_previous_page/1`, `Flop.to_next_page/2`,
   `Flop.to_previous_offset/1` or `Flop.to_next_offset/2`, and apply them to the